// src/generated/Change.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChangeOp } from "./ChangeOp";
import type { ChangeOrigin } from "./ChangeOrigin";
import type { Id } from "./Id";

/**
 * One numbered change.
 */
export type Change = { seq: number, id: Id, kind: string, op: ChangeOp, origin: ChangeOrigin, };